feat: 增加 `bsp/buzzer` 和 `bsp/flash`，增加对 `usb-cdc` 的支持，增加修改串口波特率的函数

+++ bsp/include/bsp/tim.h
@@ -7,9 +7,9 @@
 
 void bsp_tim_set(bsp_timer_handle_t *device, uint32_t period, uint32_t prescaler);
 
-void bsp_tim_config(bsp_timer_handle_t *device, double p);
+void bsp_tim_config(bsp_timer_handle_t *device, float p);
 
-void bsp_tim_set_duty(bsp_timer_handle_t *device, uint32_t channel, double duty);
+void bsp_tim_set_duty(bsp_timer_handle_t *device, uint32_t channel, float duty);
 
 void bsp_tim_pwm_enable(bsp_timer_handle_t *device, uint32_t channel);
